Signal when we remove the last element of circle
We are going to recursively call a function that takes the same position
from the circle resulting from taking that position and so on and so
forth. We need a termination condition for that recursive call. That is
when we don't have more elements to remove. When the resulting circle of
a remove operation is empty we just return a None value.